계산기 수정

+++ 책 보면서 연습!/Calculator.cpp
@@ -5,7 +5,7 @@ int Add(int a, int b)
 	return a + b;
 }
 
-float Add(int a, int b)
+float Add(float a, float b)
 {
 	return a + b;
 }
